Move exceptions to correct file
