+++ src/carbon_txt/processors.py
@@ -7,26 +7,7 @@ from arelle.RuntimeOptions import RuntimeOptions  # type: ignore
 from pydantic import BaseModel
 
 import typing
-
-
-class NoMatchingDatapointsError(ValueError):
-    """
-    Thrown when a report does not have the expected datapoint.
-
-    This could be because a data point is either
-    1. missing, or
-    2. intentionally omitted because it was deemed immaterial
-    """
-
-    def __init__(self, message: str, datapoint_short_code: str) -> None:
-        super().__init__(message)
-        self.datapoint_short_code = datapoint_short_code
-
-
-class NoLoadableCSRDFile(ValueError):
-    """
-    Thrown when a the link CSRD file can't be loaded by Arelle.
-    """
+from .exceptions import NoMatchingDatapointsError, NoLoadableCSRDFile
 
 
 class DataPoint(BaseModel):
